+++ sys/src/9/bitsy/trap.c
@@ -616,7 +616,7 @@ noted(Ureg* ureg, ulong arg0)
 
 	/* sanity clause */
 	oureg = (ulong)nureg;
-	if(!okaddr((ulong)oureg-BY2WD, BY2WD+sizeof(Ureg), 0)){
+	if(!okaddr(oureg-BY2WD, BY2WD+sizeof(Ureg), 0)){
 		qunlock(&up->debug);
 		pprint("bad ureg in noted or call to noted when not notified\n");
 		pexit("Suicide", 0);
@@ -714,7 +714,7 @@ notify(Ureg* ureg)
 	sp = ureg->sp;
 	sp -= sizeof(Ureg);
 
-	if(!okaddr((ulong)up->notify, 1, 0)
+	if(!okaddr((uintptr)up->notify, 1, 0)
 	|| !okaddr(sp-ERRMAX-4*BY2WD, sizeof(Ureg)+ERRMAX+4*BY2WD, 1)){
 		qunlock(&up->debug);
 		pprint("suicide: bad address in notify\n");
