+++ sys/src/9/omap4/trap.c
@@ -170,7 +170,7 @@ notify(Ureg *ureg)
 	}
 	sp = ureg->sp;
 	sp -= 256 + sizeof(Ureg);
-	if(!okaddr((ulong)up->notify, 1, 0)
+	if(!okaddr((uintptr)up->notify, 1, 0)
 	|| !okaddr(sp - ERRMAX - 4 * BY2WD, sizeof(Ureg) + ERRMAX + 4 * BY2WD, 1)){
 		qunlock(&up->debug);
 		pprint("suicide: bad address in notify\n");
@@ -216,7 +216,7 @@ noted(Ureg *ureg, ulong arg0)
 	
 	nureg = up->ureg;
 	oureg = (ulong) nureg;
-	if(!okaddr((ulong) oureg - BY2WD, BY2WD + sizeof(Ureg), 0)){
+	if(!okaddr(oureg - BY2WD, BY2WD + sizeof(Ureg), 0)){
 		qunlock(&up->debug);
 		pprint("bad ureg in noted or call to noted when not notified\n");
 		pexit("Suicide", 0);
